Added TSD_EU zip file with correct date
The previously-posted zip file was named TSD_EU_202010108.zip (i.e. had an extra "1" as part of the date string), this broke the link from the README.md file. I did not know how to fix the filename on github, so I downloaded the file with the wrong name, changed it, and then uploaded it again. I will delete the file with the wrong name shortly.